+++ crates/telio-lana/src/event_log_file.rs
@@ -10,7 +10,7 @@ const LOGFILE_PATH: &str = "events-moose.log";
 /// Mock for setting the device info on libmoose.
 /// Is logging every piece of info helpful or a general log would be enough?
 #[allow(unused_must_use)]
-pub fn init_device_info() {
+pub fn init_context_info() {
     let foreign_tracker = "nordvpnapp";
 
     moose::fetch_specific_context(foreign_tracker);
@@ -24,6 +24,59 @@ pub fn init_device_info() {
     event_log("set_context_device_resolution", Some(vec!["NA"]));
     event_log("set_context_device_timeZone", Some(vec!["NA"]));
     event_log("set_context_device_type", Some(vec!["NA"]));
+    event_log(
+        "set_context_user_subscription_currentState_activationDate",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_frequencyInterval",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_frequencyUnit",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_isActive",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_isNewCustomer",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_merchantID",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_paymentAmount",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_paymentCurrency",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_paymentProvider",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_paymentStatus",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_planId",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_planType",
+        Some(vec!["NA"]),
+    );
+    event_log(
+        "set_context_user_subscription_currentState_subscriptionStatus",
+        Some(vec!["NA"]),
+    );
+    event_log("set_context_user_subscription_history", Some(vec!["NA"]));
 }
 
 /// Logs a function call and its arguments to file.
@@ -121,11 +174,51 @@ pub mod moose {
         EventLogError,
     }
 
+    #[derive(Debug, Clone, Copy)]
+    pub enum MooseError {
+        Input,
+        Version,
+        StorageEngine,
+        StorageSet,
+        StorageGet,
+        History,
+        Usage,
+        Handler,
+        QueueDisabled,
+        QueueFull,
+        ContextNotFound,
+        Send,
+        NotInitiated,
+    }
+
     /// Logger result
     #[derive(Debug, PartialEq, Eq)]
     pub enum Result {
         Success,
         AlreadyInitiated,
+    }
+
+    #[derive(Debug, Clone, Copy)]
+    pub enum MooseErrorLevel {
+        Warning,
+        Error,
+    }
+
+    #[derive(Debug, Clone, Copy)]
+    pub enum ContextState {
+        Ready,
+        AlreadyInitiated,
+        TrackerDeleted,
+        AllTrackersDeleted,
+        Deleted,
+    }
+
+    pub trait InitCallback {
+        fn on_init(&self, result_code: &std::result::Result<ContextState, MooseError>);
+    }
+
+    pub trait ErrorCallback {
+        fn on_error(&self, error_level: MooseErrorLevel, error_code: MooseError, msg: &str);
     }
 
     /// Initialize logger file with current date and time.
@@ -135,7 +228,9 @@ pub mod moose {
     /// * app_name      - Lana's application name
     /// * app_version   - Semantic version of the application.
     /// * exp_moose_ver - Eventual moose version
-    /// * prod          - wether the events should be sent to production or not
+    /// * prod          - whether the events should be sent to production or not
+    /// * init_cb       - callback by Moose after succesfull initialization
+    /// * error_cb      - callback by Moose to record any errors
     #[allow(unused_variables)]
     pub fn init(
         event_path: String,
@@ -143,6 +238,8 @@ pub mod moose {
         app_version: String,
         exp_moose_ver: String,
         prod: bool,
+        init_cb: Box<(dyn InitCallback + 'static)>,
+        error_cb: Box<(dyn ErrorCallback + Sync + std::marker::Send + 'static)>,
     ) -> std::result::Result<Result, Error> {
         match super::event_log(
             "init",